+++ internal/service/artifacts/artifacts_export.go
@@ -11,6 +11,7 @@ import (
 func init() {
 	exportArtifactsContainerRepositoryHints.GetIdFn = getArtifactsContainerRepositoryId
 	exportArtifactsContainerImageSignatureHints.GetIdFn = getArtifactsContainerImageSignatureId
+	exportArtifactsGenericArtifactHints.GetIdFn = getArtifactsGenericArtifactId
 	exportArtifactsRepositoryHints.GetIdFn = getArtifactsRepositoryId
 	tf_export.RegisterCompartmentGraphs("artifacts", artifactsResourceGraph)
 }
@@ -26,8 +27,7 @@ func getArtifactsRepositoryId(resource *tf_export.OCIResource) (string, error) {
 }
 
 func getArtifactsContainerRepositoryId(resource *tf_export.OCIResource) (string, error) {
-
-	repositoryId, ok := resource.SourceAttributes["repository_id"].(string)
+	repositoryId, ok := resource.SourceAttributes["id"].(string)
 	if !ok {
 		return "", fmt.Errorf("[ERROR] unable to find repositoryId for Artifacts ContainerRepository")
 	}
@@ -43,13 +43,22 @@ func getArtifactsContainerImageSignatureId(resource *tf_export.OCIResource) (str
 	return imageSignatureId, nil
 }
 
+func getArtifactsGenericArtifactId(resource *tf_export.OCIResource) (string, error) {
+
+	artifactId, ok := resource.SourceAttributes["id"].(string)
+	if !ok {
+		return "", fmt.Errorf("[ERROR] unable to find artifactId for Artifacts GenericArtifact")
+	}
+	return artifactId, nil
+}
+
+// Hints for discovering and exporting this resource to configuration and state files
 var exportArtifactsContainerConfigurationHints = &tf_export.TerraformResourceHints{
 	ResourceClass:        "oci_artifacts_container_configuration",
 	DatasourceClass:      "oci_artifacts_container_configuration",
 	ResourceAbbreviation: "container_configuration",
 }
 
-// Hints for discovering and exporting this resource to configuration and state files
 var exportArtifactsContainerRepositoryHints = &tf_export.TerraformResourceHints{
 	ResourceClass:          "oci_artifacts_container_repository",
 	DatasourceClass:        "oci_artifacts_container_repositories",
@@ -83,10 +92,31 @@ var exportArtifactsRepositoryHints = &tf_export.TerraformResourceHints{
 	},
 }
 
+var exportArtifactsGenericArtifactHints = &tf_export.TerraformResourceHints{
+	ResourceClass:          "oci_artifacts_generic_artifact",
+	DatasourceClass:        "oci_artifacts_generic_artifacts",
+	DatasourceItemsAttr:    "generic_artifact_collection",
+	IsDatasourceCollection: true,
+	ResourceAbbreviation:   "generic_artifact",
+	RequireResourceRefresh: true,
+	DiscoverableLifecycleStates: []string{
+		string(oci_artifacts.GenericArtifactLifecycleStateAvailable),
+	},
+}
+
 var artifactsResourceGraph = tf_export.TerraformResourceGraph{
 	"oci_identity_compartment": {
+		{TerraformResourceHints: exportArtifactsContainerConfigurationHints},
 		{TerraformResourceHints: exportArtifactsContainerRepositoryHints},
 		{TerraformResourceHints: exportArtifactsContainerImageSignatureHints},
 		{TerraformResourceHints: exportArtifactsRepositoryHints},
+	},
+	"oci_artifacts_repository": {
+		{
+			TerraformResourceHints: exportArtifactsGenericArtifactHints,
+			DatasourceQueryParams: map[string]string{
+				"repository_id": "id",
+			},
+		},
 	},
 }